+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 ### Added
 - Basic VDF file syntax
 - VDF objects folding
+- VDF structure view and structure-aware navigation bar